Sign up page created and connected with backend
