+++ wiki_database_API/wikidata_django_practice/settings.py
@@ -34,7 +34,8 @@ CORS_ALLOW_ALL_ORIGINS = True
 
 ALLOWED_HOSTS = [
     config('HEROKU_APP'),
-    '127.0.0.1'
+    '127.0.0.1',
+    '10.0.2.2'
 ]
